+++ CHANGELOG.md
@@ -11,6 +11,8 @@ To know more about breaking changes, see the [Migration Guide][].
 ### Features
 
 - Support OpenHarmony.
+- [AssetPathEntity.darwinType] and [AssetPathEntity.darwinSubtype] are deprecated after v3.1.0, and use [AssetPathEntity.albumTypeEx] instead.
+
 
 ## 3.0.1